requirements.txt
tensorflow_gpu==1.1.0
scipy==1.2
matplotlib==1.5.3
numpy==1.12.1
Pillow==4.1.1
tensorflow==1.1.0
